### 2019-08-26 #1681 @whedon
[PRE REVIEW]: RHEOS - A Julia package for Rheology Data Analysis.
**Submitting author:** @akabla (<a href="http://orcid.org/0000-0002-0280-3531">Alexandre Kabla</a>)
**Repository:** <a href="https://github.com/JuliaRheology/RHEOS.jl" target ="_blank">https://github.com/JuliaRheology/RHEOS.jl</a>
**Version:** v0.9
