### 2019-08-26 #1681 @kyleniemeyer
@akabla this PR should fix the compilation issue (assuming there aren't others): https://github.com/JuliaRheology/RHEOS.jl/pull/34
