### 2019-08-26 #1681 @akabla
Dear Jed,    @jedbrown

please find below suggestions of reviewers for the paper.

@HaoZeke has most kindly offered to review.

Based on the list of JOSS reviewers, I would also suggest:

- myousefi
- wkearn
- mancellin

Many thanks
