### 2019-08-26 26c8671
fix Seifert software reference in bib file
