+++ paper/paper.bib
@@ -146,7 +146,7 @@ in the framework of tube theory and its extensions.},
 }
 
 @misc{seifertPythonToolsAnalysis2019,
-  title = {Python Tools for Analysis of {{AFM}} Data. {{Contribute}} to Jcbs/{{ForceMetric}} Development by Creating an Account on {{GitHub}}},
+  title = {Python Tools for Analysis of {{AFM}} Data.},
   copyright = {View license},
   author = {Seifert, Jacob},
   month = mar,
