### 2019-08-26 331db27
add DOI for Findley et al book
