### 2019-08-26 #1681 @moustachio-belvedere
I have removed the Findley book DOI that `whedon` suggests as it a DOI for a book review of the actual book.
