### 2019-08-26 3409c4a
remove Findley DOI, it is a just a book review
