+++ paper/paper.bib
@@ -88,7 +88,6 @@ in the framework of tube theory and its extensions.},
   isbn = {978-0-486-66016-5},
   lccn = {TA418.2 .F48 1989},
   shorttitle = {Creep and Relaxation of Nonlinear Viscoelastic Materials},
-  doi = {10.1115/1.3424077},
   language = {en},
   publisher = {{Dover}},
   author = {Findley, William N. and Lai, James S. and Onaran, Kasif},
