### 2019-08-27 #1681 @akabla
Dear @jedbrown,

This is just to check if you are waiting for more information from us.

Please let me know if there is anything else I should do in addition to the list of [suggested referees above](https://github.com/openjournals/joss-reviews/issues/1681#issuecomment-524829693).

Thanks!
