### 2019-09-02 #1700 @jedbrown
@adambeall, @gbruer15, @HaoZeke :wave: Welcome and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the [RHEOS.jl repository](https://github.com/JuliaRheology/RHEOS.jl)). I'll be watching this thread if you have any questions.
