### 2019-09-04 1328568
fix small typo in paper, industry -> industries
