### 2019-09-05 af8d64c
add spacing between badges
