### 2019-09-07 JuliaRheology/RHEOS.jl#38 @HaoZeke
[JOSS] Minor Markdown Formatting
Hi. This is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/1700. I was wondering if the image in the paper might be made a bit smaller to improve the overall pdf formatting. Specifically, at the moment, there is almost over half a page of whitespace before Figure one, as shown visually in the screenshot.
![image](https://user-images.githubusercontent.com/4336207/64467811-42c49a80-d13a-11e9-959c-a0327130dca0.png)
