### 2019-09-07 JuliaRheology/RHEOS.jl#38 @akabla
Thanks very much Rohit, I agree with you, that is not optimal. I tried to make the figure more compact, but this does not really solve the problem on its own. We need to learn a bit more about how JOSS handles figure sizes and placement. We'll investigate next week!
