### 2019-09-07 JuliaRheology/RHEOS.jl#38 @HaoZeke
Oh in that case, since I believe @whedon (and therefore JOSS) uses pandoc internally, you could play around with the [following syntax](https://pandoc.org/MANUAL.html#images):

```md
![](file.jpg){ width=50% }
```
Which essentially becomes:
```latex
\includegraphics[width=0.5\textwidth,height=\textheight]{file.jpg}
```
I'll continue with the rest of the review in the meantime.
