### 2019-09-07 cf4a2b7
figure resize test
