### 2019-09-07 74f5c41
adjust fig size and fig colour
