### 2019-09-07 JuliaRheology/RHEOS.jl#38 @HaoZeke
As of https://github.com/JuliaRheology/RHEOS.jl/commit/cf4a2b76208e24c692c3564d1281e8ef97ca3b1d
This has been resolved very nicely. Thanks!
