### 2019-09-07 JuliaRheology/RHEOS.jl#38 @akabla
Fantastic! thanks so much Rohit.
It looks much better now. I'll play a bit more with the figure size later on.
