### 2019-09-10 24d0ff6
docstrings datagen formatting update
