### 2019-09-10 68fe3e6
docstrings italics fix test
