### 2019-09-10 841a9a3
more docstrings updates
