### 2019-09-10 1c04715
more docstrings updates
