### 2019-09-10 b02714c
more docstrings updates
