### 2019-09-10 625e535
update NLOpt reference
