+++ paper/paper.bib
@@ -141,7 +141,7 @@ in the framework of tube theory and its extensions.},
 @misc{johnsonNLoptNonlinearoptimizationPackage,
   title = {The {{NLopt}} Nonlinear-Optimization Package},
   copyright = {LGPL},
-  url = {http://ab-initio.mit.edu/nlopt},
+  url = {https://github.com/stevengj/nlopt},
   author = {Johnson, Steven G}
 }
 
