### 2019-09-12 JuliaRheology/RHEOS.jl#43 @adambeall
(JOSS) Documentation of Model Limitations
Would it be possible to clarify in the paper and perhaps the documentation exactly what the requirements and limitations for a custom rheological model are? It appears that you can create any linear viscoelastic rheology as long as relaxation and / or creep moduli can be provided or fitted. The paper mentions an “emphasis on linear rheological models” and shows an example of the hereditary integral form for linear viscoelastic materials, however as someone coming from a different field, I interpreted these as options rather than strict model prerequisites. If you agree, it would only take a few added words to make this clearer.

On a related note, it may be useful to mention on the “Create Your Model” page in the documentation that it is also possible to build a model with only a creep modulus (provided only strain is predicted).
