### 2019-09-12 9b38697
fix seifert reference in BIB file
