+++ paper/paper.bib
@@ -147,8 +147,8 @@ in the framework of tube theory and its extensions.},
 
 @misc{seifertPythonToolsAnalysis2019,
   title = {Python Tools for Analysis of {{AFM}} Data.},
-  copyright = {View license},
   author = {Seifert, Jacob},
+  url = {https://github.com/jcbs/ForceMetric},
   month = mar,
   year = {2019}
 }
