### 2019-09-12 d0e9da9
raise NaN import warning only if NaNs present
