+++ test/IO.jl
@@ -1,9 +1,16 @@
-function _nanremove()
+function _nanremove_withnans()
     testarr = [1.0 1.0 NaN; NaN NaN 2.0; 3.0 3.0 3.0; NaN 4.0 4.0; 5.0 5.0 5.0]
 
     RHEOS.nanremove(testarr) == [3.0 3.0 3.0; 5.0 5.0 5.0]
 end
-@test _nanremove()
+@test _nanremove_withnans()
+
+function _nanremove_nonans()
+    testarr = [1.0 1.0 1.0; 2.0 2.0 2.0; 3.0 3.0 3.0; 4.0 4.0 4.0; 5.0 5.0 5.0]
+
+    RHEOS.nanremove(testarr) == [1.0 1.0 1.0; 2.0 2.0 2.0; 3.0 3.0 3.0; 4.0 4.0 4.0; 5.0 5.0 5.0]
+end
+@test _nanremove_nonans()
 
 function _importcsv_t_withnans()
     fildir = joinpath(@__DIR__, "testdata", "datawithnans.csv")
