### 2019-09-13 #1700 @adambeall
Thank you for the opportunity to review. I have enjoyed trying out RHEOS and think that it is accessible, the examples hosted in the repository nicely demonstrate its functionality and the documentation is thorough. RHEOS should be useful for researchers from a range of disciplines teaching and forward modelling a variety of linear viscoelastic rheologies, as well as fitting experimental data (the Bonfanti et al., 2019 demonstration is impressive).

Thanks to the authors for making the minor recommended changes. While [issue 43](https://github.com/JuliaRheology/RHEOS.jl/issues/43) is still open, I will be unavailable for the next week and don’t want to hold up the review, so I have completed the checklist and recommend acceptance after the issue is closed.
