### 2019-09-13 JuliaRheology/RHEOS.jl#43 @akabla
Thanks very much Adam @adambeall , this is a fair point indeed.
You are right that at the moment linearity is a constraint of the methods implemented, although we have some plans to extend this in future.
We will address this properly next week.
