### 2019-09-13 #1700 @akabla
Thank you so much Adam for your support and all your input and comments throughout. We will address issue 43 very soon - you made a good point indeed, and it should be straight-forward to address.
