### 2019-09-14 2281d21
nanremove syntax neatened up
