### 2019-09-14 1cd80aa
even neater nanremove syntax
