### 2019-09-15 JuliaRheology/RHEOS.jl#46 @HaoZeke
[JOSS] References
Hi,
Very minor issue brought up by @whedon here, https://github.com/openjournals/joss-reviews/issues/1700#issuecomment-531591218

Would be happy to close once it has been addressed.
![image](https://user-images.githubusercontent.com/4336207/64926448-32d44700-d81b-11e9-9b8c-ff4d415450d0.png)
