### 2019-09-15 JuliaRheology/RHEOS.jl#46 @akabla
Thanks so much Rohit @HaoZeke for checking.
Whedon is actually picking up a [review of the book](https://asmedigitalcollection.asme.org/appliedmechanics/article/44/2/364/388377/Creep-and-Relaxation-of-Nonlinear-Viscoelastic), and not the book itself. I can't find a doi for the book, and I'm not sure there is one. We'll look again.
