### 2019-09-15 JuliaRheology/RHEOS.jl#46 @moustachio-belvedere
Thanks for noting that @HaoZeke. Indeed, as @akabla mentioned, I can confirm that the DOI is a review of the book.

I did not notice it was a review at first and added in the DOI to the reference but Whedon still did not accept it so I looked a bit more closely. (This was during the pre-review https://github.com/openjournals/joss-reviews/issues/1681 , commit 331db27b80f47a6631fe4a952363ca9c766b7831 added it and then commit https://github.com/JuliaRheology/RHEOS.jl/commit/3409c4afa0e68d5b97ef4f4c8a884893ea3ca544 removed it once I had realised). I should have realised before as it is unusual for a book to have a DOI as far as I understand.

I will close the issue for now but please do let us know if you have any further queries on that.

Thanks again
