### 2019-09-15 JuliaRheology/RHEOS.jl#46 @HaoZeke
An thanks for clearing this up! Sorry I hadn't noticed it in the pre review.
