### 2019-09-15 JuliaRheology/RHEOS.jl#46 @moustachio-belvedere
No worries at all,
