### 2019-09-17 #1700 @gbruer15
I've finished my review and recommend accepting this submission, pending issue [#43](https://github.com/JuliaRheology/RHEOS.jl/issues/43). Other than that issue, the software's documentation is in a good, user-friendly state, and I believe this package is a very useful contribution to rheology.
