### 2019-09-17 #1700 @HaoZeke
I have completed my review @jedbrown, and as soon as https://github.com/JuliaRheology/RHEOS.jl/issues/43 is closed satisfactorily, I recommend this for publication. RHEOS is very well documented and easily extensible, it's really well structured and addresses an existing gap in the rheology community. Glad to have been introduced to it, @akabla.
