### 2019-09-17 f111401
clarify limitation to linear models in paper and doc
