### 2019-09-17 2b67a91
Small changes create your model notebook
