### 2019-09-17 JuliaRheology/RHEOS.jl#43 @akabla
I have clarified the limitations of the current limitation of the library to the class of linear models, in the paper as well as in the [doc intro](https://juliarheology.github.io/RHEOS.jl/latest/).

Adam's comment regarding the "create your own model" section to follow.
