### 2019-09-17 41cfc64
Change notebook creat your own model
