### 2019-09-17 JuliaRheology/RHEOS.jl#43 @ab2425
Thanks a lot Adam @adambeall. We have addressed the linearity limitation in the "create your own model" as well as the possibility of defining a model with only one modulus.

The following commits should have resolved this issue:
f1114016ece9ebd35d75711d294785f78aa24e0c
2b67a912a59e31e5827d593cd2fff4b1ff2ebc98
41cfc646d34685379e38ff193d9d1ea5cd624f65

Please let us know if you have further queries.
