### 2019-09-17 #1700 @moustachio-belvedere
Many thanks from me too @adambeall @gbruer15 @HaoZeke for all your valuable feedback and being so efficient in working through everything. We are very grateful.
