### 2019-09-17 5af67cc
feature add: semi-specified init conditions and boundaries, no documentation yet
