+++ src/processing.jl
@@ -174,8 +174,86 @@ end
 Fitting and predicting functions
 --------------------------------
 =#
+function fill_init_params(model::RheoModelClass, p0::Union{NamedTuple,Nothing})
+
+    springpotnumber = length(findall(x->(x==:a)||(x==:β), model.params))
+
+    if isnothing(p0)
+        # fill all initial guesses as 0.5, if two springpots then
+        # make sure dissipation coefficients conform to constraints
+        p0a = ones(RheoFloat, length(model.params))/2
+        if springpotnumber==2
+            index = findfirst(model.params.==:a)
+            p0a[index] = RheoFloat(0.8)
+        end
+        @warn "Initial values for model parameters are set to $p0a by default"
+
+    elseif length(p0)<length(model.params)
+        # check all provided symbols are part of the model, fill all parameters not provided
+        for i in keys(p0)
+            @assert (i in model.params) "Incorrect parameter name provided for model initial conditions:" * i
+        end
+        # propagate list
+        p0a = Vector{RheoFloat}(undef, length(model.params))
+        for (i,p) in enumerate(model.params)
+            p0a[i] = (p in keys(p0)) ? p0[p] : RheoFloat(1/2)
+        end
+        # check for second springpot specification
+        indexa = findfirst(model.params.==:a)
+        indexβ = findfirst(model.params.==:β)
+        if !(:a in keys(p0)) && !(:β in keys(p0))
+            p0a[indexa] = RheoFloat(0.8)
+        elseif !(:β in keys(p0))
+            p0a[indexβ] = p0[:a]/2
+        elseif !(:a in keys(p0))
+            p0a[indexa] = minimum((p0[:β]*2, RheoFloat(0.999)))
+        end
+
+        @warn "Unspecified initial guesses filled by default values: $p0a"
+
+    else
+        p0a = model_parameters(p0, model.params, "initial guess")
+    end
+
+    return p0a
+end
+
+function fill_lower_bounds(model::RheoModelClass, lo::Union{NamedTuple,Nothing})
+    if isnothing(lo)
+        loa = nothing
+
+    elseif length(lo)<length(model.params)
+        loa = Vector{RheoFloat}(undef, length(model.params))
+        for (i,p) in enumerate(model.params)
+            loa[i] = (p in keys(lo)) ? lo[p] : zero(RheoFloat)
+        end
+
+    else
+        loa = model_parameters(lo, model.params, "low bounds")
+    end
+
+    return loa
+end
+
+function fill_upper_bounds(model::RheoModelClass, hi::Union{NamedTuple,Nothing})
+    if isnothing(hi)
+        hia = nothing
+    
+    elseif length(hi)<length(model.params)
+        hia = Vector{RheoFloat}(undef, length(model.params))
+        for (i,p) in enumerate(model.params)
+            hia[i] = (p in keys(hi)) ? hi[p] : RheoFloat(Inf)
+        end
+
+    else
+        hia = model_parameters(hi, model.params, "high bounds")
+    end
+
+    return hia
+end
+
 """
-    modelfit(data::RheoTimeData, model::RheoModelClass, modloading::Symbol; p0::Union{NamedTuple,Tuple} = (), lo::Union{NamedTuple,Tuple} = (), hi::Union{NamedTuple,Tuple} = (), verbose::Bool = false, rel_tol = 1e-4, diff_method="BD")
[…]
