### 2019-09-18 e1ead2f
array{rheofloat,1} to vector{rheofloat} in definitions.jl
