### 2019-09-18 b679a4b
add gamma constraint for fractional zener
