### 2019-09-18 193401a
add gamma constraint for fractional poynting-thompson
