### 2019-09-18 780ab9e
move p0 assertion into filling function
