### 2019-09-18 a61d055
singularity offset magic number removed
