### 2019-09-18 b3652ff
fix fill_init tests for added assertion
