### 2019-09-18 5bd77fa
fix modelsteppredict approximation for singularity
