### 2019-09-18 3d819db
add convenience function to check modulus is defined without confusing magic number
