### 2019-09-18 07042db
add check that modulus is defined for dynamicmodel fit
