+++ src/processing.jl
@@ -711,6 +711,12 @@ function dynamicmodelfit(data::RheoFreqData,
     loa = fill_lower_bounds(model, lo)
     hia = fill_upper_bounds(model, hi)
 
+    # check necessary moduli are defined
+    modsingGp = (ω->model.Gp(ω,p0a))
+    modsingGpp = (ω->model.Gpp(ω,p0a))
+    @assert modulusexists(modsingGp) "Storage modulus not defined for this model"
+    @assert modulusexists(modsingGpp) "Loss modulus not defined for this model"
+
     # initialise NLOpt.Opt object with :LN_SBPLX Subplex algorithm
     opt = Opt(:LN_SBPLX, length(p0a))
 
