### 2019-09-18 8a7a625
JOSS paper: add reviewers to acknowledgements
