### 2019-09-18 54edb1d
JOSS paper: fix typo in acknowledgements
