### 2019-09-22 b0e2db6
Minor copy edits, APA bib style/capitalization, and serial commas
