+++ paper/paper.md
@@ -27,9 +27,9 @@ bibliography: paper.bib
 
 Rheology is the science of deformation and flow, with a focus on materials that do not exhibit simple linear elastic or viscous Newtonian behaviours. Rheology plays an important role in the characterisation of soft viscoelastic materials commonly found in the food and cosmetics industries, as well as in biology and bioengineering. Empirical and theoretical approaches are commonly used to identify and quantify material behaviours based on experimental data.
 
-RHEOS (RHEology, Open-Source) is a software package designed to make the analysis of rheological data simpler, faster and more reproducible. RHEOS is currently limited to the broad family of linear viscoelastic models. A particular strength of the library is its ability to handle rheological models containing fractional derivatives, which have demonstrable utility for the modelling of biological materials [@aime2018power;@bouzid2018computing;@bonfantiUnifiedRheologicalModel2019; @kaplanPectinMethylesterificationImplications2019] but have hitherto remained in relative obscurity -- possibly due to their mathematical and computational complexity. RHEOS is written in Julia [@bezansonJuliaFreshApproach2017], which provides excellent computational efficiency and approachable syntax. RHEOS is fully documented and has extensive testing coverage.
+RHEOS (RHEology, Open-Source) is a software package designed to make the analysis of rheological data simpler, faster, and more reproducible. RHEOS is currently limited to the broad family of linear viscoelastic models. A particular strength of the library is its ability to handle rheological models containing fractional derivatives, which have demonstrable utility for the modelling of biological materials [@aime2018power;@bouzid2018computing;@bonfantiUnifiedRheologicalModel2019; @kaplanPectinMethylesterificationImplications2019], but have hitherto remained in relative obscurity -- possibly due to their mathematical and computational complexity. RHEOS is written in Julia [@bezansonJuliaFreshApproach2017], which provides excellent computational efficiency and approachable syntax. RHEOS is fully documented and has extensive testing coverage.
 
-To our knowledge there is to this date no other software package that offers RHEOS' broad selection of rheology analysis tools and extensive library of both traditional and fractional models. It has been used to process data and validate a model in [@bonfantiUnifiedRheologicalModel2019], and is currently in use for several ongoing projects.
+To our knowledge, there is to this date no other software package that offers RHEOS' broad selection of rheology analysis tools and extensive library of both traditional and fractional models. It has been used to process data and validate a model in @bonfantiUnifiedRheologicalModel2019, and is currently in use for several ongoing projects.
 
 It should be noted that RHEOS is not an optimisation package. It builds on another optimisation package, NLopt [@johnsonNLoptNonlinearoptimizationPackage], by adding a large number of abstractions and functionality specific to the exploration of viscoelastic data.
 
@@ -37,7 +37,7 @@ It should be noted that RHEOS is not an optimisation package. It builds on anoth
 
 ## Arbitrary stress-strain curves and broad relaxation spectra require advanced software
 
-Many scientists and engineers who undertake rheological experiments would fit their data with one or several viscoelastic models in order to classify materials, quantify their behaviour and predict their response to external perturbations.
+Many scientists and engineers who undertake rheological experiments would fit their data with one or several viscoelastic models in order to classify materials, quantify their behaviour, and predict their response to external perturbations.
 
[…]
