### 2019-09-22 #1700 @jedbrown
Thanks for your patience.
* Would it make sense to name the authors of Bob-rheology? https://sourceforge.net/p/bob-rheology/wiki/Home/  There are three authors listed in the copyright statements inside the repository, with version 2.5 release in 2012. Perhaps we can name the three authors and use `year = 2012`. I've suggested that in my PR https://github.com/JuliaRheology/RHEOS.jl/pull/49
* Where does the year 1989 come from for [this book](https://store.doverpublications.com/0486660168.html )? It looks to be a [1976 book](https://watermark.silverchair.com/364_1.pdf?token=AQECAHi208BE49Ooan9kkhW_Ercy7Dm3ZL_9Cf3qfKAc485ysgAAAmwwggJoBgkqhkiG9w0BBwagggJZMIICVQIBADCCAk4GCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMKW5euSTlIsgNAvRZAgEQgIICH-dgrsCFgUgCwFr6n9BevlIHKG1UKGyLXfwsoVjfONgNmdfmfQjUxLTpUfJQrwMN0eSaq3Xr_7iliGvzlQnuBjmZf2K5RV8VWwtIa6CQEn5irVdUvcsXEtoI55pmZKEDZ5FJacZhTml7svEe25eRJPDhcB-D0c27WOMIQPwzmXpbMjhEo1M4bAZNNQwAz580I6Y9XbxehGmoHk9J0Z1TE_tVuObGLTnneLaOmHWRHUzdzUz8NK7-9B0-n_S40mgZHY9OsF3tFMksBIFhjsnMY_3-XrO9TewMVxiWyCOb3FLc390XYhVEJwQNLa9fBJsOdQe0P0DHSxU9NkHnR3RwSBKwySBX6iGLrIDlZgCsG2MgnKhDcvhsE0yqT9TVzmsbhMZhn1im9Mwi2hb31k3jXjSnzIW1APlQCZF_Oi5_TfWs5bJueJSvY3zZPH3AcALGfNns3oknHzgXrNK1eu5ndqKV1MeVWe3jz9n2OhVD7lntDd1ff5HMyEOQNMAMkDOOaJsujAkmgtqJgc9_AOiJV7pnBGo30FMkfMbCuTFr6HCPvbVh6JPEakLdjbsFtTY9K_ClIuAhuRVlHD01N7_-64_SWv-t9LW6nzDQGwjQae-Fm3CXnABA_QsfTwSyt23wclrQySSCNApmSkm_A6b6kgijwd6S4rth3wSW8ZYETh3m1NDP0tDlGNzofdZj_GEbAMqHIMNyE7UNDEZrSn9jFg) reprinted in 2011. If the year needs to change, the sentence in which it's mentioned will have to be updated as well.
* Perhaps you could rephrase that sentence in which you name many textbooks to just say "over the past N decades" because the APA-style citations include the year anyway.
