### 2019-09-22 #1700 @moustachio-belvedere
@jedbrown Many thanks for your feedback. On your first point I agree completely, thanks a lot for the pull request.

On the third point I also agree completely, I hadn't noticed the referencing style and have updated the wording in line with your suggestion, commit https://github.com/JuliaRheology/RHEOS.jl/commit/2186c1d9efabf3f69e009780cb051c061ca65681

Regarding the book reference, it does indeed come from the Dover publications version. However, I believe their first reprint was in 1989. The front matter of the version I have states the following:

_Copyright @ 1976 by North-Holland Publishing Company.
All rights reserved.

This Dover edition, first published in 1989, is an unabridged, corrected
republication of the work originally published in 1976 by North-Holland
Publishing Company, Amsterdam, as Volume 18 in the “North-Holland
Series in Applied Mathematics and Mechanics.” A new section of additional
literature relating to Chapter 2 has been added in this edition at the end of the
bibliography._

I would also note that the 1989 version is sometimes cited, e.g. in the popular MIT lecture notes:

http://web.mit.edu/course/3/3.11/www/modules/visco.pdf

As I don't know exactly how much was abridged in the original 1976 version I might be hesitant to cite that edition. But if there are referencing precedents that the first edition should be cited then I am more than happy to oblige. I also very happy to check with our department librarian if there are precedents that they are aware of regarding this.

Many thanks again.
