+++ paper/paper.bib
@@ -95,6 +95,7 @@ in the framework of tube theory and its extensions.},
   author = {Findley, William N. and Lai, James S. and Onaran, Kasif},
   year = {1989},
   keywords = {Viscoelasticity,Creep,Materials,Stress relaxation (Physics)},
+  origdate = {1976}
 }
 
 @article{kaplanPectinMethylesterificationImplications2019,
