### 2019-09-22 #1700 @moustachio-belvedere
Another update, for now  I tried to add an `origdate` field with `1976` embedded but it does not seem to be recognised by the compiler backend pipeline, perhaps there is another field which could be used. The addition was based on this discussion:

https://tex.stackexchange.com/questions/139805/how-to-differentiate-between-translated-and-reprinted-work-with-apa-style

Many thanks
