+++ paper/paper.bib
@@ -96,6 +96,7 @@ in the framework of tube theory and its extensions.},
   year = {1989},
   keywords = {Viscoelasticity,Creep,Materials,Stress relaxation (Physics)},
   origdate = {1976},
+  note = {Original work published 1976},
 }
 
 @article{kaplanPectinMethylesterificationImplications2019,
