### 2019-09-22 #1700 @jedbrown
I tried following [this example](https://github.com/jgm/pandoc-citeproc/blob/master/tests/biblatex-examples.bib#L809-L836) from `pandoc-citeproc` (the software pipeline we use), but those fields apparently are not supported, so I'm fine with this as is.

Please tag a release ([annotated tag](https://git-scm.com/book/en/v2/Git-Basics-Tagging#_annotated_tags) preferred) and archive your software on Zenodo or similar, then report the DOI back here.
