### 2019-09-22 62993b4
clean-up definitions.jl - union operator overload
