### 2019-09-22 13f3414
clean-up definitions.jl - union operator overload
