### 2019-09-22 3acd0cc
clean-up definitions.jl - minor formatting
