### 2019-09-22 a4321f5
clean-up definitions.jl - RheoModelClass and other formatting
