### 2019-09-22 28a7368
clean-up definitions.jl - log documentation
