### 2019-09-22 2bfb4d1
fix Findley reference date in docs and README in light of JOSS discussion
