### 2019-09-22 d5f9839
add sketch of RheoTimeData union operator
