### 2019-09-22 ab68a88
add sketch of RheoTimeData union operator
