### 2019-09-22 a815ba0
add sketch of RheoTimeData union operator
