### 2019-09-22 37c2e26
add tests for RheoTimeData union operator
