+++ test/definitions.jl
@@ -21,16 +21,25 @@ end
 
 function _union_strain1stress2()
     time_instance = timeline(t_start=0.0, t_end=15.0, step=0.2)
-    imposed_strain = stressfunction(time_instance, sawtooth(offset=5.0, amp=2, period=5))
+    imposed_strain = strainfunction(time_instance, sawtooth(offset=5.0, amp=2, period=5))
     imposed_stress = stressfunction(time_instance, stairs(offset=3.0, amp=0.5, width=4.0))
 
-    println(imposed_strain.log)
-    
     combined = imposed_strain|imposed_stress
 
     (combined.σ==imposed_stress.σ) && (combined.ϵ==imposed_strain.ϵ) && (combined.t==imposed_stress.t)
 end
 @test _union_strain1stress2()
+
+function _union_stress1strain2()
+    time_instance = timeline(t_start=0.0, t_end=15.0, step=0.2)
+    imposed_strain = strainfunction(time_instance, sawtooth(offset=5.0, amp=2, period=5))
+    imposed_stress = stressfunction(time_instance, stairs(offset=3.0, amp=0.5, width=4.0))
+
+    combined = imposed_stress|imposed_strain
+
+    (combined.σ==imposed_stress.σ) && (combined.ϵ==imposed_strain.ϵ) && (combined.t==imposed_stress.t)
+end
+@test _union_stress1strain2()
 
 function _freeze_params()
     SLS2_mod = freeze_params( SLS2, (G₀=2,η₂=3.5))
