### 2019-09-23 #1700 @jedbrown
@akabla Please edit the author info on Zenodo to match the paper. (Kyle and myself should not be authors.) Looks good otherwise and we're ready to accept.
