### 2019-09-24 315a804
update modelsteppredict to contain Inf and warning
