### 2019-09-24 d39b5bb
small update to modelsteppredict docstring
