### 2019-09-24 8e78c72
fix broken link in index.md to contributors
