### 2019-09-24 #1700 @jedbrown
@openjournals/joss-eics This paper is now accepted; over to you.
