### 2019-09-24 #1700 @moustachio-belvedere
Thanks very much for editing @jedbrown .

And thankyou @danielskatz for finalizing the process.

The entire review process has been very valuable to us.
